fix rpc::cluster::new_node
git-svn-id: svn+ssh://code.etolabo.org/usr/etolabo/var/svn/ficia/branches/frsyuki/kumofs@334 b300df90-00d7-44b1-9271-3fa3ef26909a

+++ logic/mgr.cc
@@ -80,7 +80,7 @@ void Manager::new_node(address addr, role_type id, shared_node n)
 		return;
 
 	} else {
-		LOG_ERROR("unkown node id ",id);
+		LOG_ERROR("unkown node id ",(uint16_t)id);
 	}
 }
